requirements.txt
flet
python-telegram-bot
requests
beautifulsoup4
yt-dlp
google-generativeai